+++ modi/build.gradle
@@ -31,6 +31,8 @@ subprojects {
         implementation 'org.springframework.boot:spring-boot-starter-actuator'
         implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 
+        implementation 'org.springdoc:springdoc-openapi-starter-webmvc-ui:2.6.0'
+
         implementation 'org.springframework.boot:spring-boot-starter-security'
         implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
         implementation 'io.jsonwebtoken:jjwt-impl:0.11.5'
