+++ modi/common/src/main/java/com/coc/modi/common/auth/JwtTokenProvider.java
@@ -1,9 +1,8 @@
-package com.coc.modi.auth.infrastructure.jwt;
+package com.coc.modi.common.auth;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
-import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
