+++ modi/member-service/src/main/java/com/coc/modi/auth/application/MemberAuthService.java
@@ -3,7 +3,7 @@ package com.coc.modi.auth.application;
 import com.coc.modi.auth.application.dto.MemberLoginCommand;
 import com.coc.modi.auth.application.dto.MemberLoginResponse;
 import com.coc.modi.auth.presentation.dto.MemberLoginInfo;
-import com.coc.modi.auth.infrastructure.jwt.JwtTokenProvider;
+import com.coc.modi.common.auth.JwtTokenProvider;
 import com.coc.modi.member.domain.Member;
 import com.coc.modi.member.infrastructure.MemberRepository;
 import lombok.RequiredArgsConstructor;
