+++ modi/member-service/src/main/java/com/coc/modi/member/MemberServiceApplication.java
@@ -9,7 +9,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
-@SpringBootApplication(scanBasePackages = "com.coc.modi")
+@SpringBootApplication(scanBasePackages = {"com.coc.modi"})
 @EnableJpaAuditing
 @SecurityScheme(
         name = "BearerAuth",
